🗃️ ftree: rename schedule folder to schedules
