+++ src/ui/global/widgets/layouts/private-layout/header/header-title/index.tsx
@@ -1,7 +1,9 @@
+'use client'
+
 import { useHeaderTitle } from './use-header-title'
 
 export function HeaderTitle() {
   const { pageTitle } = useHeaderTitle()
 
-  return <h1 className='text-2xl font-extrabold text-black md:text-3xl'>{pageTitle}</h1>
+  return <h1 className='text-2xl font-extrabold text-black'>{pageTitle}</h1>
 }
