+++ src/ui/global/widgets/layouts/private-layout/header/index.tsx
@@ -1,25 +1,24 @@
-import { Button } from "@heroui/button";
+import { Button } from '@heroui/button'
 
-import { HeaderTitle } from "./header-title";
-import { Icon } from "../../../components/Icon";
+import { HeaderTitle } from './header-title'
+import { Icon } from '../../../components/Icon'
 
 type HeaderProps = {
-  title?: string;
-  onMenuClick: () => void;
-};
+  onMenuClick: () => void
+}
 
-export function Header({ onMenuClick, title = "Dashboard" }: HeaderProps) {
+export function Header({ onMenuClick }: HeaderProps) {
   return (
-    <header className="flex items-center px-6 h-24 bg-white border-b border-gray-border">
+    <header className='flex items-center px-6 h-16 bg-white border-b border-gray-border'>
       <Button
         onPress={onMenuClick}
-        variant="ghost"
-        size="sm"
-        className="p-0 mr-4 text-white border-none bg-blue-primary md:hidden"
+        variant='ghost'
+        size='sm'
+        className='p-0 mr-4 text-white border-none bg-blue-primary md:hidden'
       >
-        <Icon name="menu" size={24} />
+        <Icon name='menu' size={24} />
       </Button>
       <HeaderTitle />
     </header>
-  );
+  )
 }
